Refatoração das Classes Account

+++ src/main/java/com/wipro/techbank/services/SpecialAccountService.java
@@ -1,10 +1,15 @@
 package com.wipro.techbank.services;
 
+import com.wipro.techbank.domain.Client;
 import com.wipro.techbank.domain.SpecialAccount;
-import com.wipro.techbank.dtos.SpecialAccountDto;
+import com.wipro.techbank.dtos.SpecialAccountRequestDto;
+import com.wipro.techbank.dtos.SpecialAccountResponseDto;
+import com.wipro.techbank.repositories.ClientRepository;
 import com.wipro.techbank.repositories.SpecialAccountRepository;
 import com.wipro.techbank.services.exceptions.DataBasesException;
 import com.wipro.techbank.services.exceptions.ResourceNotFoundException;
+import org.modelmapper.ModelMapper;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -12,61 +17,69 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
 @Service
 public class SpecialAccountService {
 
     @Autowired
-    SpecialAccountRepository specialAccountRepository;
+    private SpecialAccountRepository specialAccountRepository;
+    @Autowired
+    private ClientRepository clientRepository;
 
-    // public SpecialAccount save(SpecialAccount specialAccount){ return specialAccountRepository.save(specialAccount);}
+    @Autowired
+    private ModelMapper modelMapper;
 
-    public SpecialAccountDto create(SpecialAccountDto dto) {
-        SpecialAccount entity = new SpecialAccount();
-        copyDtoToEntity(dto, entity);
-        entity = specialAccountRepository.save(entity);
-        return new SpecialAccountDto(entity);
+    public List<SpecialAccountResponseDto> findAll(){
+        return specialAccountRepository.findAll()
+                .stream()
+                .map(this::toSpecialAccountDto)
+                .collect(Collectors.toList());
     }
 
-    public Page<SpecialAccountDto> findAllPaged(Pageable pageable) {
-        Page<SpecialAccount> specialAccounts = specialAccountRepository.findAll(pageable);
-        return specialAccounts.map(SpecialAccountDto::new);
+    public SpecialAccountResponseDto findById(Long id){
+        Optional<SpecialAccount> optionalSpecialAccount = specialAccountRepository.findById(id);
+        SpecialAccount specialAccount = optionalSpecialAccount.orElseThrow(() ->
+                new ResourceNotFoundException("Entidade não encontrada"));
+        return toSpecialAccountDto(specialAccount);
     }
 
-    public SpecialAccountDto findById(Long id){
-        SpecialAccount specialAccount = specialAccountRepository.findById(id).get();
-        SpecialAccountDto dto = new SpecialAccountDto();
-        return copyDtoToDTO(dto, specialAccount);
+    public void create(SpecialAccountRequestDto specialAccountRequestDto){
+        Long idClient = specialAccountRequestDto.getClient().getId();
+        Optional<Client> client = clientRepository.findById(idClient);
+        if(!client.isPresent()){
+            throw new ResourceNotFoundException("Entidade não encontrada");
+        }
+        SpecialAccount specialAccount = toSpecialAccount(specialAccountRequestDto);
+        specialAccountRepository.save(specialAccount);
     }
 
-    public SpecialAccount update(Long id, SpecialAccount specialAccount){
-        SpecialAccount account = specialAccountRepository.findById(id).orElseGet(() ->{
-            throw new ResourceNotFoundException("Conta não encontrada");
-        });
-        return specialAccountRepository.save(account);
+    public SpecialAccountResponseDto updateSpecialAccount(Long id, SpecialAccountRequestDto specialAccountRequestDto){
+        Optional<SpecialAccount> optionalSpecialAccountDb = specialAccountRepository.findById(id);
+        SpecialAccount specialAccountDb = specialAccountRepository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Entidade não econtrada."));
+
+        BeanUtils.copyProperties(specialAccountRequestDto,specialAccountDb);
+        specialAccountRepository.save(specialAccountDb);
+        return toSpecialAccountDto(specialAccountDb);
     }
 
-    public void delete(Long id){
-        try {
+    public void remove(Long id){
+        Optional<SpecialAccount> optionalSpecialAccount = specialAccountRepository.findById(id);
+        if(optionalSpecialAccount.isPresent()){
             specialAccountRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new ResourceNotFoundException("Id " + id + " não encontrado.");
-        } catch (DataIntegrityViolationException e) {
-            throw new DataBasesException("Violação de integridade.");
+        } else {
+            throw new ResourceNotFoundException("Entidade não encontrada");
         }
     }
 
-    private void copyDtoToEntity(SpecialAccountDto dto, SpecialAccount entity) {
-        entity.setId(dto.getId());
-        entity.setBalance(dto.getBalance());
-        entity.setCreditSpecial(dto.getCreditSpecial());
-        entity.setCreditSpecialUsed(dto.getCreditSpecialUsed());
+    private SpecialAccountResponseDto toSpecialAccountDto(SpecialAccount specialAccount){
+        return modelMapper.map(specialAccount, SpecialAccountResponseDto.class);
     }
 
-    private SpecialAccountDto copyDtoToDTO(SpecialAccountDto dto, SpecialAccount entity) {
-        dto.setId(entity.getId());
-        dto.setBalance(entity.getBalance());
-        dto.setCreditSpecial(entity.getCreditSpecial());
-        dto.setCreditSpecialUsed(entity.getCreditSpecialUsed());
-        return dto;
+    private SpecialAccount toSpecialAccount(SpecialAccountRequestDto specialAccountRequestDto){
+        return modelMapper.map(specialAccountRequestDto, SpecialAccount.class);
     }
 }
